Mavenize, Add in 1.8 dependencies
Start to upgrade for 1.8. It seems like it should work anyway, But we'll
have to give it a few test runs and see if there are some bugs.
