+++ .gitignore
@@ -1,8 +1,25 @@
-*.class
+.DS_Store*
+ehthumbs.db
+Icon?
+Thumbs.db
+bin
+target
+*.glob
+.*.glob
 .classpath
 .project
-.settings/
+.settings
+.idea/
+*.iml
+*.ipr
+*.iws
+dependency-reduced-pom.xml
+/target/
+*.class
+.project
+.classpath
+/.settings/
 
-# Package Files #
-*.war
-*.ear
+/lib/
+.settings
+.git
