+++ src/main/resources/config.yml
@@ -1,0 +1,93 @@
+#
+# Config for dsTitle plugin
+# For the different text colors and styles, go to
+# http://www.minecraftwiki.net/wiki/Formatting_codes
+# Be aware to use spaces, not tabs.
+# For more info, see https://github.com/denniss17/dsTitles/blob/master/dsTitle/config.yml
+#
+general:
+  # Should the plugin set the chatformat to the one specified in this config?
+  overwrite_format: false
+  # The charformat used when overwrite_format is true
+  # %1$s will be replaced with the playername, %2$s with the message and
+  # [titleprefix] and [titlesuffix] with the title
+  chat_format: '<[titleprefix]%1$s[titlesuffix]> %2$s'
+  # The tag in the chatformat which will be replaced by the prefix
+  chat_format_prefix_tag: '[titleprefix]'
+  # The tag in the chatformat which will be replaced by the suffix
+  chat_format_suffix_tag: '[titlesuffix]'
+  # If chatformat is not overwritten, but the prefixtag or suffixtag cannot be found,
+  # The prefix of suffix will be inserted before or after this tag in the chatformat, 
+  # which will most of the times be the playersname.
+  # By default this is '%1$s', but some chat plugins use other tags for the playername
+  chat_format_player_tag: '%1$s'
+  # Set this to true if you use ChatManager
+  use_deprecated_listener: false
+  # Should the plugin show all unavailable titles when a player does /title list ?
+  show_unavailable_titles: true
+  # Should chat title be enabled?
+  use_chattag: true
+  # Should the plugin also change the tags above the heads of players?
+  # These will be set to the values of headprefix and headsuffix in titleConfig.yml
+  # Disable this if you have another plugin which handles this or uses teams of the Bukkit scoreboard api.
+  use_nametag: true
+  # Should the plugin check for updates and notify all players with ds_title.admin?
+  check_for_updates: true
+  # The time between two checks for a new version, in minutes (default 180 = 3 hours)
+  update_check_interval: 180
+storage:
+  # Storage type to store titles of player
+  # Possible values: 'file' and 'database'
+  type: file
+  # The name of the file to use when storing titles in a file
+  file: players.yml
+  # Database settings
+  database:
+    # Database driver. Possible values: H2, MYSQL, POSTGRE, SQLITE
+    driver: ''
+    # Url of the database, or filename in case of SQLITE
+    url: ''
+    # Username for database. Not needed when using SQLITE
+    username: ''
+    # Password for database. Not needed when using SQLITE
+    password: ''
+messages:
+  # I think this doesn't need explanation
+  error_no_player: '&8[&eTitles&8] &cOnly players can use this command'
+  error_no_permission: '&8[&eTitles&8] &cYou don`t have enough permissions for this command'
+  error_no_vault: '&8[&eTitles&8] &cVault needs to be installed/enabled to use this command'
+  error_no_valid_field: '&8[&eTitles&8] &cThat is not a valid field. Valid fields: &fchattag&c, &fheadtag&c, &fpermission&c, &fdescription'
+  error_prefix_not_found: '&8[&eTitles&8] &cThat prefix is not found!'
+  error_suffix_not_found: '&8[&eTitles&8] &cThat suffix is not found!'
+  error_prefix_exists: '&8[&eTitles&8] &cThat prefix already exists!'
+  error_suffix_exists: '&8[&eTitles&8] &cThat suffix already exists!'
+  error_headtag_too_long: '&8[&eTitles&8] &cA headtag can have a maximum length of 16!'
+  prefix_set: '&8[&eTitles&8] &ePrefix set!'
+  suffix_set: '&8[&eTitles&8] &eSuffix set!'
+  title_cleared: '&8[&eTitles&8] &eTitle cleared!'
+  title_added: '&8[&eTitles&8] &eTitle added!'
+  title_edited: '&8[&eTitles&8] &eTitle edited!'
+  prefix_granted: '&8[&eTitles&8] &ePrefix &f{title} &egranted to &f{name}&e.'
+  suffix_granted: '&8[&eTitles&8] &eSuffix &f{title} &egranted to &f{name}&e.'
+  prefix_ungranted: '&8[&eTitles&8] &ePrefix &f{title} &eungranted from &f{name}&e.'
+  suffix_ungranted: '&8[&eTitles&8] &eSuffix &f{title} &eungranted from &f{name}&e.'
+  reloaded: '&8[&eTitles&8] &eConfig reloaded'
+  update_notification: '&8[&eTitles&8] &eThere is a new version (&f{version}&e, current: &f{current}&e) available. Download it here: {website}'
+  # Header showed above unavailable titles, when a player does /title list
+  unavailable_header: '&8~~=={  &cUnavailable &etitles &8}==~~'
+  # Header showed above available titles, when a player does /title list
+  available_header: '&8~~=={  &aAvailable &etitles &8}==~~'
+  # Make up of one title in the list showed when a player does /title list
+  title_listitem: '{name}&r: {preview} &r&8({description})'
+  # The help menu (if you just use /title)
+  menu_header: '&8~~=={ &e&ldsTitles help &r&8}==~~'
+  menu_prefix_list: ' &e/title prefix &fList available prefixes'
+  menu_prefix_set: ' &e/title prefix set <name> &fSet your prefix'
+  menu_suffix_list: ' &e/title suffix &fList available suffixes'
+  menu_suffix_set: ' &e/title suffix set <name> &fSet your suffix'
+  menu_clear: ' &e/title clear &fClear your title'
+  menu_add: ' &e/title add <suffix|prefix> <title> &fAdd a new title'
+  menu_edit: ' &e/title edit <suffix|prefix> <title> <field> <value> &fEdit a field of a title'
+  menu_grant: ' &e/title grant <suffix|prefix> <player> <title> &fGrant a title to a player'
+  menu_ungrant: ' &e/title ungrant <suffix|prefix> <player> <title> &fUngrant a title from a player'
+  menu_reload: ' &e/title reload &fReload the config'
