+++ src/main/resources/titleConfig.yml
@@ -1,0 +1,27 @@
+#
+# This is the config file where you put the title
+# These have to be in the following format:
+# prefixes:
+#   <name>:
+#     chattag: <chattag>
+#     headtag: <headtag>
+#     permission: <permission needed>
+#     description: <description>
+# suffixes:
+#   <name>:
+#     chattag: <chattag>
+#     headtag: <headtag>
+#     permission: <permission needed>
+#     description: <description>
+# default_prefix: <name>
+# default_suffix: <name>
+#
+# Chattag, headtag, permission and description are optional, although it is usefull
+# if at least one of chattag and headtag is set ;-)
+#
+#
+#
+prefixes: {}
+suffixes: {}
+default_prefix: ''
+default_suffix: ''
